Add ${COMPILER_SETTINGS} to ${MINGW_COMMAND}
Build systems (such as Meson/Autotools) may make incorrect assumptions about the build environment which are corrected by ${COMPILER_SETTINGS} but it needed to be manually specified. Since ${MINGW_COMMAND} is used for target environment setup, ${COMPILER_SETTINGS} has now been added to it for ensuring that the compiler env is correct.